day1/02-prompting-basics/bedrock/requirements.txt
boto3>=1.42.0
python-dotenv>=1.0.0
gradio>=5.0.0
streamlit>=1.40.0
